% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplplotapi.R
\name{assayAnnotations}
\alias{assayAnnotations}
\title{Get a list of fields available for x- and y-axis of assay annotation plot}
\arguments{
\item{api_key}{String - the api key for the CTX API}

\item{url_env}{String of environment that api should be called in}
}
\description{
Get a list of fields available for x- and y-axis of assay annotation plot
}
