% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_assay_descriptions}
\alias{available_assay_descriptions}
\title{A list of all available assays that have an associated pdf description}
\format{
full_assay_list
vector of integer values of aeids that have associated data
}
\source{
invitrodb
}
\usage{
available_assay_descriptions
}
\description{
This is a list of aeids that have an associated descriptive pdf stored in clowder
}
\keyword{datasets}
