% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{cnst}
\alias{cnst}
\title{Title}
\usage{
cnst(ps, x)
}
\arguments{
\item{ps}{Unused in cnst function}

\item{x}{Log base 10 concentration}
}
\value{
response value at given concentration based on function parameters
}
\description{
Title
}
