% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{exp3}
\alias{exp3}
\title{exp3}
\usage{
exp3(a, b, p, x)
}
\arguments{
\item{a}{Function Parameter}

\item{b}{Function Parameter}

\item{p}{Function Parameter}

\item{x}{Log base 10 concentration}
}
\value{
response value at given concentration based on function parameters
}
\description{
exp3
}
