% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{exp4}
\alias{exp4}
\title{Exp 4}
\usage{
exp4(tp, ga, x)
}
\arguments{
\item{tp}{model top}

\item{ga}{Function Parameter}

\item{x}{Log base 10 concentration}
}
\value{
response value at given concentration based on function parameters
}
\description{
Exp 4
}
