% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{exp5}
\alias{exp5}
\title{Exp5}
\usage{
exp5(tp, ga, p, x)
}
\arguments{
\item{tp}{Function Parameter}

\item{ga}{Function Parameter}

\item{p}{Function Parameter}

\item{x}{Log base 10 concentration}
}
\value{
response value at given concentration based on function parameters
}
\description{
Exp5
}
