% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_assay_list}
\alias{full_assay_list}
\title{A list of all available assay endpoints saved as an rda file}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1570 rows and 48 columns.
}
\source{
https://ccte-api-ccd.epa.gov/ccdapp2/assay-endpoint-detail/
}
\usage{
full_assay_list
}
\description{
This dataset is used so that we don't have to call a
large api endpoint multiple times everytime someone loads a plot
}
\keyword{datasets}
