% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{gnls_curve}
\alias{gnls_curve}
\title{Gain-Loss function}
\usage{
gnls_curve(top, ga, gw, la, lw, lconc)
}
\arguments{
\item{top}{model top}

\item{ga}{gnls Gain AC50}

\item{gw}{Hill coefficient in the gain direction}

\item{la}{AC50 in the loss direction for the gain-loss model}

\item{lw}{Hill coefficient in the loss direction}

\item{lconc}{Log base 10 concentration}
}
\value{
response value at given concentration based on function parameters
}
\description{
Gain-Loss function
}
