% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{hill_curve}
\alias{hill_curve}
\title{Hill Function}
\usage{
hill_curve(hill_tp, hill_ga, hill_gw, lconc)
}
\arguments{
\item{hill_tp}{Top parametre for hill model}

\item{hill_ga}{Gain ac50}

\item{hill_gw}{Hill coefficient in the gain direction}

\item{lconc}{Log base 10 concentration}
}
\value{
response value at given concentration based on function parameters
}
\description{
Hill Function
}
