% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplplotapi.R
\name{plot}
\alias{plot}
\title{Plot out data from invitrodb}
\arguments{
\item{dsstox_id}{String of the dsstox substance id}

\item{assay_endpoint_nm}{String of the assay endpoint name}

\item{representative_sample}{TRUE/FALSE should only the representative sample be displayed}

\item{url_env}{String of environment that api should be called in}
}
\description{
Plot out data from invitrodb
}
