% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplplotapi.R
\name{plotAssays}
\alias{plotAssays}
\title{Plot out assay annotations in matrix format}
\arguments{
\item{xaxis}{Character, the x-axis field name from dat, default "intendedTargetFamily"}

\item{yaxis}{Character, the y-axis field name from dat, default "organism"}

\item{api_key}{String - the api key for the CTX API}

\item{url_env}{String of environment that api should be called in}
}
\description{
Plot out assay annotations in matrix format
}
