% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplplotapi.R
\name{plotfit2}
\alias{plotfit2}
\title{Plot out data from httr/htpp}
\arguments{
\item{cellType}{String of the cell type for the chemical}

\item{signature}{String of the signature name}

\item{sampleId}{String of the sample Id that should be plotted}

\item{dataset}{String of the dataset that the sample is included in}

\item{url_env}{String of environment that api should be called in}
}
\description{
Plot out data from httr/htpp
}
