% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplplotapi.R
\name{plotfit2_HTPP}
\alias{plotfit2_HTPP}
\title{Plot out data from htpp conc resp}
\arguments{
\item{sampleId}{String of the sample Id that should be plotted}

\item{categoryName}{String of the categoryName that the sample is included in}

\item{url_env}{String of environment that api should be called in}
}
\description{
Plot out data from htpp conc resp
}
