% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplplotapi.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot Summary data for dtxsid}
\arguments{
\item{dsstox_id}{String - the dsstox substance id}

\item{url_env}{String of environment that api should be called in}
}
\description{
Plot Summary data for dtxsid
}
