% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplplotapi.R
\name{plotSummaryHTPP}
\alias{plotSummaryHTPP}
\title{Plot Summary for HTPP Summary data}
\arguments{
\item{dsstox_id}{String - the dsstox substance id}

\item{approach}{String - either "category" or "feature" determines which subset to plot}

\item{url_env}{String - environment that api should be called in}
}
\description{
Plot Summary for HTPP Summary data
}
