% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplplotapi.R
\name{plotSummaryHTTr}
\alias{plotSummaryHTTr}
\title{Plot Summary for HTTr Summary data}
\arguments{
\item{dsstox_id}{String - the dsstox substance id}

\item{url_env}{String of environment that api should be called in}
}
\description{
Plot Summary for HTTr Summary data
}
