% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{poly1}
\alias{poly1}
\title{Poly1}
\usage{
poly1(a, x)
}
\arguments{
\item{a}{Function Parameter}

\item{x}{Log base 10 concentration}
}
\value{
response value at given concentration based on function parameters
}
\description{
Poly1
}
