% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{poly2}
\alias{poly2}
\title{Poly2}
\usage{
poly2(a, b, x)
}
\arguments{
\item{a}{Function Parameter}

\item{b}{Function Parameter}

\item{x}{Log base 10 concentration}
}
\value{
response value at given concentration based on function parameters
}
\description{
Poly2
}
