% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{pow}
\alias{pow}
\title{Pow}
\usage{
pow(a, p, x)
}
\arguments{
\item{a}{Function Parameter}

\item{p}{Function Parameter}

\item{x}{Log base 10 concentration}
}
\value{
response value at given concentration based on function parameters
}
\description{
Pow
}
