% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotSummary.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Function for switching to scientific notation in ccd}
\usage{
specify_decimal(x, n = 3)
}
\arguments{
\item{x}{Number to be displayed}

\item{n}{Number of decimal places to round to}
}
\description{
Function for switching to scientific notation in ccd
}
