% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplApiQuery.R
\name{tcplApiQuery}
\alias{tcplApiQuery}
\title{tcplApiQuery used to get upstream API data for bioactivity (invitrodb) data}
\usage{
tcplApiQuery(
  dsstox_sid,
  aenm,
  rs = TRUE,
  url_env = "https://ccte-api-ccd-dev.epa.gov/"
)
}
\arguments{
\item{dsstox_sid}{String - dsstox substance id of the form DTXSIDXXXXXX}

\item{aenm}{String - the name of the assay tested. Must match name in data hub exactly}

\item{rs}{Bool - default TRUE do we want to return only the representative sample.}

\item{url_env}{String - the environment in which the api should hit.}
}
\value{
json
}
\description{
tcplApiQuery used to get upstream API data for bioactivity (invitrodb) data
}
