% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplApiQuery.R
\name{tcplAssayAnnotationsLoad}
\alias{tcplAssayAnnotationsLoad}
\title{tcplAssayAnnotationsLoad used to load assay annotations from CTX API}
\usage{
tcplAssayAnnotationsLoad(
  api_key,
  url_env = "https://api-ccte.epa.gov/bioactivity"
)
}
\arguments{
\item{api_key}{String - the api key for the CTX API}

\item{url_env}{String - the environment in which the api should hit.}
}
\value{
assays a data frame containing all assays and valid annotations for matrix
}
\description{
tcplAssayAnnotationsLoad used to load assay annotations from CTX API
}
