% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplfit2Plot.R
\name{tcplfit2Plot}
\alias{tcplfit2Plot}
\title{Plotting for data related to tcplfit2 (Httr, HTPP)}
\usage{
tcplfit2Plot(dat)
}
\arguments{
\item{dat}{single row dataframe loaded from api for a single signature}
}
\value{
plotly figure
}
\description{
Plotting for data related to tcplfit2 (Httr, HTPP)
}
