% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplfit2Plot.R
\name{tcplfit2PlotHTPP}
\alias{tcplfit2PlotHTPP}
\title{Plotting for data related to tcplfit2 (HTPP)}
\usage{
tcplfit2PlotHTPP(dat)
}
\arguments{
\item{dat}{single row dataframe loaded from api for a single signature}
}
\value{
plotly figure
}
\description{
Plotting for data related to tcplfit2 (HTPP)
}
