% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplApiQuery.R
\name{tcplHTPPConcResp}
\alias{tcplHTPPConcResp}
\title{API call to GET HTPP Conc/Resp Data}
\usage{
tcplHTPPConcResp(
  sampleId,
  categoryName,
  url_env = "https://ccte-api-ccd-dev.epa.gov/"
)
}
\arguments{
\item{sampleId}{String Sample ID}

\item{categoryName}{String associated category name for htpp data}

\item{url_env}{String URL base where the api should call the data from}
}
\value{
res_dt a datatable with all of the httr conc/resp data collapsed
}
\description{
API call to GET HTPP Conc/Resp Data
}
