% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplApiQuery.R
\name{tcplHTTrConcResp}
\alias{tcplHTTrConcResp}
\title{API call to GET HTTr Conc/Resp Data}
\usage{
tcplHTTrConcResp(
  cellType,
  signature,
  sampleId,
  dataset,
  url_env = "https://ccte-api-ccd-dev.epa.gov/"
)
}
\arguments{
\item{cellType}{String the celltype}

\item{signature}{String signature for the associated data}

\item{sampleId}{String sample id otherwise known as spid}

\item{dataset}{String required dataset indicator for httr}

\item{url_env}{String url where the api should be called}
}
\value{
res_dt a datatable with all of the httr conc/resp data collapsed
}
\description{
API call to GET HTTr Conc/Resp Data
}
