% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotAssays.R
\name{tcplPlotAssays}
\alias{tcplPlotAssays}
\title{Bioactivity assay annotations clustering matrix plot}
\usage{
tcplPlotAssays(assays, xaxis = "intendedTargetFamily", yaxis = "organism")
}
\arguments{
\item{assays}{dataframe with assay annotations from output of tcplAssayAnnotationsLoad}

\item{xaxis}{Character, the x-axis field name from 'assays'}

\item{yaxis}{Character, the y-axis field name from 'assays'}
}
\value{
Annotations plot
}
\description{
Bioactivity assay annotations clustering matrix plot
}
