% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotlyMultiPlot.R
\name{tcplPlotlyMultiPlot}
\alias{tcplPlotlyMultiPlot}
\title{Function for plotting multiple plots at the same time}
\usage{
tcplPlotlyMultiPlot(dat)
}
\arguments{
\item{dat}{a dataframe with each plot's associated parameters per row}
}
\value{
A plot or multiple plots in a single html widget
}
\description{
Function for plotting multiple plots at the same time
}
