% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotlyPlot.R
\name{tcplPlotlyPlot}
\alias{tcplPlotlyPlot}
\title{tcplPlotlyPlot a generic plotly plotting function that accepts JSON as input}
\usage{
tcplPlotlyPlot(dat)
}
\arguments{
\item{dat}{dataframe of necessary parameters for plotting}
}
\value{
generic plotly plot of given JSON parameters
}
\description{
tcplPlotlyPlot a generic plotly plotting function that accepts JSON as input
}
