% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotSummary.R
\name{tcplPlotSummary}
\alias{tcplPlotSummary}
\title{Summary plot for bioactivity (invitrodb) data}
\usage{
tcplPlotSummary(dat = tcplPlot::tcplSummaryPlotExampleData)
}
\arguments{
\item{dat}{dataframe with plotting parameters from output of tcplApiQuery}
}
\value{
Summary level plot
}
\description{
Summary plot for bioactivity (invitrodb) data
}
