% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotSummary.R
\name{tcplPlotSummaryAEID}
\alias{tcplPlotSummaryAEID}
\title{Summary plot for bioactivity (invitrodb) data}
\usage{
tcplPlotSummaryAEID(dat)
}
\arguments{
\item{dat}{dataframe with plotting parameters from output of tcplApiQuery}
}
\value{
Summary level plot
}
\description{
Summary plot for bioactivity (invitrodb) data
}
