% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotSummary.R
\name{tcplPlotSummaryHTPP}
\alias{tcplPlotSummaryHTPP}
\title{Summary Plot for HTPP data}
\usage{
tcplPlotSummaryHTPP(dat, approach = "category")
}
\arguments{
\item{dat}{dataframe with all associated data required to plot}

\item{approach}{either category or feature to indicate which subset to filter}
}
\value{
Summary level plot of the HTPP data
}
\description{
Summary Plot for HTPP data
}
