% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplPlotSummary.R
\name{tcplPlotSummaryHTTr}
\alias{tcplPlotSummaryHTTr}
\title{Summary Plot for HTTr data}
\usage{
tcplPlotSummaryHTTr(dat)
}
\arguments{
\item{dat}{dataframe with all associated plotting data as columns}
}
\value{
Plot with combined checkbox functionality
}
\description{
Summary Plot for HTTr data
}
