% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplApiQuery.R
\name{tcplSummaryHTPPLoad}
\alias{tcplSummaryHTPPLoad}
\title{API call to GET HTPP Summary Data}
\usage{
tcplSummaryHTPPLoad(dsstox_sid, url_env = "https://ccte-api-ccd-dev.epa.gov/")
}
\arguments{
\item{dsstox_sid}{dsstox substance id}

\item{url_env}{url environment that the api should call}
}
\value{
res_dt a datatable with all of the httr summary data collapsed
}
\description{
API call to GET HTPP Summary Data
}
