% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplApiQuery.R
\name{tcplSummaryHTTrLoad}
\alias{tcplSummaryHTTrLoad}
\title{API call to GET HTTr Summary Data}
\usage{
tcplSummaryHTTrLoad(dsstox_sid, url_env = "https://ccte-api-ccd-dev.epa.gov/")
}
\arguments{
\item{dsstox_sid}{string dsstox substance id}

\item{url_env}{string url of environment that the data should be pulled from}
}
\value{
res_dt a datatable with all of the httr summary data collapsed
}
\description{
API call to GET HTTr Summary Data
}
