% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplApiQuery.R
\name{tcplSummaryLoad}
\alias{tcplSummaryLoad}
\title{API query to load summary level bioactivity (invitrodb) data}
\usage{
tcplSummaryLoad(dsstox_sid, url_env = "https://ccte-api-ccd-dev.epa.gov/")
}
\arguments{
\item{dsstox_sid}{String dsstox substance id}

\item{url_env}{String url environment where the upstream api should be queried}
}
\value{
m5_res a dataframe with all of the summary level bioactivity data
}
\description{
API query to load summary level bioactivity (invitrodb) data
}
