% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcplSummaryPlotExampleData}
\alias{tcplSummaryPlotExampleData}
\title{Example data that can be used in place of an api call}
\format{
tcplSummaryPlotExampleData
dataframe 223 obs. of 23 variables
}
\source{
upstream api call
}
\usage{
tcplSummaryPlotExampleData
}
\description{
This is a dataframe of summary level data that can be used in place
of an upstream api call
}
\keyword{datasets}
